move rpc server
